skip udev rules if udev adm not available

+++ src/__xtz/setup_ledger.lua
@@ -6,30 +6,34 @@ ami_assert(type(_user) == "string", "User not specified...", EXIT_INVALID_CONFIG
 local platform = am.app.get_model("PLATFORM")
 
 if not _options["skip-udev"] or platform ~= "unix" then
-	-- Setup udev rules
-	local _ok, _userPlugin = am.plugin.safe_get("user")
-	if not _ok then
-		log_error("Failed to load user plugin!")
-		return
-	end
-	ami_assert(_userPlugin.add_into_group(_user, "plugdev"), "Failed to add user '" .. _user .. "' to plugdev")
+	if not os.execute("which udevadm") then
+		log_warn("udevadm not found. Skipping udev setup...")
+	else
+		-- Setup udev rules
+		local _ok, _userPlugin = am.plugin.safe_get("user")
+		if not _ok then
+			log_error("Failed to load user plugin!")
+			return
+		end
+		ami_assert(_userPlugin.add_into_group(_user, "plugdev"), "Failed to add user '" .. _user .. "' to plugdev")
 
-	local _tmpFile = os.tmpname()
-	local _udevRulesUrl =
-	"https://raw.githubusercontent.com/alis-is/udev-rules/f15dc1eb83a4f3c666f58c12a93c45c6fca3a004/add_udev_rules.sh"
-	local _ok, _error = net.safe_download_file(_udevRulesUrl, _tmpFile, { followRedirects = true })
-	if not _ok then
+		local _tmpFile = os.tmpname()
+		local _udevRulesUrl =
+		"https://raw.githubusercontent.com/alis-is/udev-rules/f15dc1eb83a4f3c666f58c12a93c45c6fca3a004/add_udev_rules.sh"
+		local _ok, _error = net.safe_download_file(_udevRulesUrl, _tmpFile, { followRedirects = true })
+		if not _ok then
+			fs.remove(_tmpFile)
+			ami_error("Failed to download: " .. tostring(_error))
+		end
+		local _proc = proc.spawn("/bin/bash", { _tmpFile }, {
+			stdio = { stderr = "pipe" },
+			wait = true,
+			env = { HOME = path.combine(os.cwd(), "data") }
+		})
+
 		fs.remove(_tmpFile)
-		ami_error("Failed to download: " .. tostring(_error))
+		ami_assert(_proc.exitcode == 0, "Failed to setup udev rules : " .. _proc.stderrStream:read("a"))
 	end
-	local _proc = proc.spawn("/bin/bash", { _tmpFile }, {
-		stdio = { stderr = "pipe" },
-		wait = true,
-		env = { HOME = path.combine(os.cwd(), "data") }
-	})
-
-	fs.remove(_tmpFile)
-	ami_assert(_proc.exitcode == 0, "Failed to setup udev rules : " .. _proc.stderrStream:read("a"))
 end
 
 local serviceManager = require "__xtz.service-manager"
